Moved the SQLite wrapper to a separate file
